+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2024-05-16
+
+### Added
+
+- Added SiteLink information to the Domain and Trust diagram
+
+### Changed
+
+- Increased Diagrammer.Core minimum version requirement v0.2.1
+
+### Fixed
+
+- Fix [#19](https://github.com/rebelinux/Diagrammer.Microsoft.AD/issues/19)
+
 ## [0.2.2] - 2024-05-08
 
 ### Added